+++ src/main/java/com/ssafy/ssashinsa/heyfy/account/controller/AccountControllerTest.java
@@ -1,8 +1,7 @@
 package com.ssafy.ssashinsa.heyfy.account.controller;
 
-import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.auth.AccountAuthResponseDto;
-import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.history.InquireTransactionHistoryResponseDto;
 import com.ssafy.ssashinsa.heyfy.account.service.AccountService;
+import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.auth.AccountAuthResponseDto;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -24,11 +23,4 @@ public class AccountControllerTest { // 테스트용 컨트롤러
         return ResponseEntity.ok(accounts);
     }
 
-    @GetMapping("/transactionhistorytest")
-    public ResponseEntity<InquireTransactionHistoryResponseDto> getTransactionHistoryTest() {
-
-        InquireTransactionHistoryResponseDto transactionHistory = accountService.getTransactionHistory();
-
-        return ResponseEntity.ok(transactionHistory);
-    }
 }
