+++ src/main/java/com/ssafy/ssashinsa/heyfy/inquire/controller/InquireController.java
@@ -1,16 +1,21 @@
 package com.ssafy.ssashinsa.heyfy.inquire.controller;
 
+import com.ssafy.ssashinsa.heyfy.inquire.docs.GetForeignTransactionHistoryDocs;
+import com.ssafy.ssashinsa.heyfy.inquire.docs.GetTransactionHistoryDocs;
 import com.ssafy.ssashinsa.heyfy.account.dto.AccountNoDto;
-import com.ssafy.ssashinsa.heyfy.inquire.docs.CheckAccountDocs;
+import com.ssafy.ssashinsa.heyfy.inquire.dto.TransactionHistoryDto;
+import com.ssafy.ssashinsa.heyfy.inquire.dto.TransactionHistoryResponseRecDto;
 import com.ssafy.ssashinsa.heyfy.inquire.docs.InquireDepositListDocs;
 import com.ssafy.ssashinsa.heyfy.inquire.docs.InquireSingleDepositDocs;
 import com.ssafy.ssashinsa.heyfy.inquire.docs.InquireSingleForeignDepositDocs;
-import com.ssafy.ssashinsa.heyfy.inquire.dto.*;
+import com.ssafy.ssashinsa.heyfy.inquire.dto.ForeignSingleDepositResponseDto;
+import com.ssafy.ssashinsa.heyfy.inquire.dto.SingleDepositResponseDto;
 import com.ssafy.ssashinsa.heyfy.inquire.service.InquireService;
+import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.history.InquireTransactionHistoryResponseDto;
+import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.history.InquireTransactionHistoryResponseRecDto;
 import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.inquire.ShinhanInquireDepositResponseDto;
 import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.inquire.ShinhanInquireDepositResponseRecDto;
 import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.inquire.ShinhanInquireSingleDepositResponseDto;
-import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -18,8 +23,9 @@ import org.springframework.web.bind.annotation.*;
 
 import java.text.DecimalFormat;
 import java.util.List;
+import java.util.stream.Collectors;
 
-@Tag(name = "예금주 계좌 조회", description = "신한은행 계좌 정보 조회 관련 API")
+@Tag(name = "계좌/거래내역 조회", description = "신한은행 계좌 정보 조회 관련 API")
 @RestController
 @RequestMapping("/inquire")
 @RequiredArgsConstructor
@@ -77,31 +83,67 @@ public class InquireController {
         return ResponseEntity.ok(finalResponse);
     }
 
-    @Hidden
-    @GetMapping("/singledeposittest") // 백엔드 테스트용 엔드포인트
-    public ResponseEntity<ShinhanInquireSingleDepositResponseDto> inquireSingleDepositTest() {
 
-        ShinhanInquireSingleDepositResponseDto response = inquireService.inquireSingleDeposit();
 
-        return ResponseEntity.ok(response);
+    @GetTransactionHistoryDocs
+    @PostMapping("/transactionhistory")
+    public ResponseEntity<TransactionHistoryResponseRecDto> getTransactionHistoryTest(@RequestBody AccountNoDto accountNo) {
+
+        InquireTransactionHistoryResponseDto originalResponseDto = inquireService.getTransactionHistory(accountNo.getAccountNo());
+        InquireTransactionHistoryResponseRecDto originalRec = originalResponseDto.getREC();
+
+        List<TransactionHistoryDto> newList = originalRec.getList().stream()
+                .map(originalItem -> TransactionHistoryDto.builder()
+                        .transactionUniqueNo(originalItem.getTransactionUniqueNo())
+                        .transactionDate(originalItem.getTransactionDate())
+                        .transactionTime(originalItem.getTransactionTime())
+                        .transactionType(originalItem.getTransactionType())
+                        .transactionTypeName(originalItem.getTransactionTypeName())
+                        .transactionAccountNo(originalItem.getTransactionAccountNo())
+                        .transactionBalance(String.valueOf((int) originalItem.getTransactionBalance()))
+                        .transactionAfterBalance(String.valueOf((int) originalItem.getTransactionAfterBalance()))
+                        .transactionSummary(originalItem.getTransactionSummary())
+                        .transactionMemo(originalItem.getTransactionMemo())
+                        .build())
+                .collect(Collectors.toList());
+
+        TransactionHistoryResponseRecDto finalResponse = TransactionHistoryResponseRecDto.builder()
+                .totalCount(originalRec.getTotalCount())
+                .list(newList)
+                .build();
+
+        return ResponseEntity.ok(finalResponse);
     }
 
-    @Hidden
-    @GetMapping("/singleforeigndeposittest") // 백엔드 테스트용 엔드포인트
-    public ResponseEntity<ShinhanInquireSingleDepositResponseDto> inquireForeignSingleDepositTest() {
+    @GetForeignTransactionHistoryDocs
+    @PostMapping("/foreigntransactionhistory")
+    public ResponseEntity<TransactionHistoryResponseRecDto> getForeignTransactionHistoryTest(@RequestBody AccountNoDto accountNo) {
 
-        ShinhanInquireSingleDepositResponseDto response = inquireService.inquireSingleForeignDeposit();
+        InquireTransactionHistoryResponseDto originalResponseDto = inquireService.getForeignTransactionHistory(accountNo.getAccountNo());
+        InquireTransactionHistoryResponseRecDto originalRec = originalResponseDto.getREC();
 
-        return ResponseEntity.ok(response);
-    }
+        DecimalFormat df = new DecimalFormat("0.00");
+        List<TransactionHistoryDto> newList = originalRec.getList().stream()
+                .map(originalItem -> TransactionHistoryDto.builder()
+                        .transactionUniqueNo(originalItem.getTransactionUniqueNo())
+                        .transactionDate(originalItem.getTransactionDate())
+                        .transactionTime(originalItem.getTransactionTime())
+                        .transactionType(originalItem.getTransactionType())
+                        .transactionTypeName(originalItem.getTransactionTypeName())
+                        .transactionAccountNo(originalItem.getTransactionAccountNo())
+                        .transactionBalance(df.format(originalItem.getTransactionBalance()))
+                        .transactionAfterBalance(df.format(originalItem.getTransactionAfterBalance()))
+                        .transactionSummary(originalItem.getTransactionSummary())
+                        .transactionMemo(originalItem.getTransactionMemo())
+                        .build())
+                .collect(Collectors.toList());
 
-    @CheckAccountDocs
-    @GetMapping("/accountcheck")
-    public ResponseEntity<AccountCheckDto> checkAccount() {
-        boolean isAccountCheck = inquireService.checkAccount();
-        AccountCheckDto response = new AccountCheckDto(isAccountCheck);
+        TransactionHistoryResponseRecDto finalResponse = TransactionHistoryResponseRecDto.builder()
+                .totalCount(originalRec.getTotalCount())
+                .list(newList)
+                .build();
 
-        return ResponseEntity.ok(response);
+        return ResponseEntity.ok(finalResponse);
     }
