+++ src/main/java/com/ssafy/ssashinsa/heyfy/inquire/docs/GetForeignTransactionHistoryDocs.java
@@ -1,7 +1,7 @@
-package com.ssafy.ssashinsa.heyfy.account.docs;
+package com.ssafy.ssashinsa.heyfy.inquire.docs;
 
 import com.ssafy.ssashinsa.heyfy.account.dto.AccountNoDto;
-import com.ssafy.ssashinsa.heyfy.account.dto.TransactionHistoryResponseRecDto;
+import com.ssafy.ssashinsa.heyfy.inquire.dto.TransactionHistoryResponseRecDto;
 import com.ssafy.ssashinsa.heyfy.common.exception.ErrorResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
