+++ src/main/java/com/ssafy/ssashinsa/heyfy/inquire/docs/GetTransactionHistoryDocs.java
@@ -1,8 +1,8 @@
-package com.ssafy.ssashinsa.heyfy.account.docs;
+package com.ssafy.ssashinsa.heyfy.inquire.docs;
 
 
 import com.ssafy.ssashinsa.heyfy.account.dto.AccountNoDto;
-import com.ssafy.ssashinsa.heyfy.account.dto.TransactionHistoryResponseRecDto; // 수정된 응답 DTO 임포트
+import com.ssafy.ssashinsa.heyfy.inquire.dto.TransactionHistoryResponseRecDto; // 수정된 응답 DTO 임포트
 import com.ssafy.ssashinsa.heyfy.common.exception.ErrorResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
